Stricter prevention of duplicates
In some cases, multiple duplicates of urls can be inserted into the search database, resulting in duplicates of search results. This change makes eliminating duplicate entires more robust. Content that generates the same md5 hash will now be treated as equal regardless of the URL associated with it, and the longer URL of the two will be discarded.

While this fix removes duplicates after the fact, it would probably be better to prevent them in the first place.