Fix troubles with InnoDB
InnoDB is supposed to support TRUNCATE TABLE when there are no foreign key constraints on the table, but since I continue to find duplicated URIs in the entry list, I suspect it's occasionally tripping over this MySQL statement. Change the TRUNCATE query to DROPping the table altogether and RENAME-ing the temp table.